Use separate caching for NVD checks
The NVD database should be cached independently of the dependencies of
the project. This stores the NVD database under a key based on the
data of running the check. If that key isn't found, it will fetch
whatever is the latest NVD database key in the cache.

Note that if no cache is found in the current branch, it will be
looked up in the main branch.